Remove substring tests from ssh rules.
The test creates a condition in which the file is not valid from SSH's
perspective causing the Ansible remediation to always fail because it
runs the ssh validation command. If the file has incorrect syntax then
there's nothing we can do other than expect that the administrator fix
the file first before applying the remediation.