OCP4: The flowcontrol API was bumped to v1beta2 in 4.11
The flowcontrol.apiserver.k8s.io API was bumped from v1beta1 to v1beta2
in upstream kube 1.24 which is available in OCP 4.11.

To support that, a new versioned rule was added, which checks the beta2
API. The old versionless rule was renamed so that it's clear that it's
only checking the beta1 API.

In addition, a new versionless api_server_api_priority_flowschema_catch_all
that OR-s the versioned ones so that in the future we can just keep
adding new versions to it without exposing more rules to our users.